requirements.txt
appdirs==1.4.3
Django==1.10.6
django-ckeditor==5.2.1
django-colorfield==0.1.12
django-filebrowser-no-grappelli==3.7.0
django-filer==1.2.7
django-gravatar2==1.4.0
django-hitcount==1.2.2
django-jet==1.0.5
django-mathfilters==0.4.0
django-meta==1.3.2
django-mptt==0.8.7
django-polymorphic==1.0.2
easy-thumbnails==2.3
feedparser==5.2.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
python-monkey-business==1.0.0
six==1.10.0
Unidecode==0.4.20
